+++ basalt-server-lib/src/repositories/users.rs
@@ -9,78 +9,16 @@ use sqlx::prelude::FromRow;
 use sqlx::SqliteExecutor;
 use utoipa::ToSchema;
 
-#[derive(
-    Debug, Clone, Copy, Hash, Eq, PartialEq, Serialize, Deserialize, utoipa::ToSchema, sqlx::Type,
-)]
-#[repr(i32)]
-#[serde(rename_all = "kebab-case")]
-pub enum Role {
-    Competitor = 0,
-    Host = 1,
-}
+use crate::{define_id_type, define_sqlx_enum};
 
-impl From<i32> for Role {
-    fn from(value: i32) -> Self {
-        match value {
-            1 => Role::Host,
-            _ => Role::Competitor,
-        }
+define_sqlx_enum! {
+    pub enum Role {
+        Competitor,
+        Host,
     }
 }
 
-impl From<i64> for Role {
-    fn from(value: i64) -> Self {
-        match value {
-            1 => Role::Host,
-            _ => Role::Competitor,
-        }
-    }
-}
-
-impl From<Role> for i32 {
-    fn from(value: Role) -> Self {
-        match value {
-            Role::Competitor => 0,
-            Role::Host => 1,
-        }
-    }
-}
-
-#[derive(
-    Debug,
-    Clone,
-    Hash,
-    Eq,
-    PartialEq,
-    Serialize,
-    Deserialize,
-    ToSchema,
-    derive_more::From,
-    derive_more::Into,
-    sqlx::Type,
-)]
-#[sqlx(transparent)]
-pub struct UserId(pub String);
-
-impl UserId {
-    // Default feels wrong here as each call to this function generates a different value.
-    #[allow(clippy::new_without_default)]
-    pub fn new() -> Self {
-        use rand::{distributions::Alphanumeric, Rng};
-        let id = rand::thread_rng()
-            .sample_iter(Alphanumeric)
-            .take(20)
-            .map(char::from)
-            .collect::<String>();
-        Self(id)
-    }
-}
-
-impl Display for UserId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
+define_id_type!(UserId);
 
 #[derive(Debug, Clone, Hash, Eq, PartialEq, FromRow, Serialize, Deserialize, ToSchema)]
 #[serde(rename_all = "camelCase")]
